chore: upgrade to nextsj 11
