+++ ggircs-app/app/next-env.d.ts
@@ -1,60 +1,6 @@
-import { NextComponentType, NextPageContext } from "next";
-import { GraphQLTaggedNode, MutationConfig } from "relay-runtime";
-import { NextRouter } from "next/router";
-import { ComponentClass } from "react";
-import { CacheConfig } from "react-relay-network-modern/node8";
+/// <reference types="next" />
+/// <reference types="next/types/global" />
+/// <reference types="next/image-types/global" />
 
-interface PageInitialProps {
-  pageProps: {
-    router: NextRouter;
-    variables: Record<string, any>;
-  };
-}
-
-interface PageComponentProps {
-  query?: any;
-  router?: NextRouter;
-}
-
-export type PageComponent = NextComponentType<
-  NextPageContext,
-  PageInitialProps,
-  PageComponentProps
-> &
-  ComponentClass<PageComponentProps> & {
-    query: GraphQLTaggedNode;
-    isAccessProtected: boolean;
-    allowedGroups: string[];
-  };
-
-export interface CacheConfigWithDebounce extends CacheConfig {
-  debounceKey?: string;
-}
-
-export interface MutationConfigWithDebounce<T> extends MutationConfig<T> {
-  cacheConfig?: CacheConfigWithDebounce;
-}
-
-export interface EcccFile {
-  name: string;
-  size: number;
-  created_at: string;
-}
-
-export interface RelayReportObject {
-  node: {
-    id: string;
-    swrsReportId: number;
-  };
-}
-
-export interface SwrsReportData {
-  submissionDate: string;
-  ecccXmlFileByEcccXmlFileId: {
-    xmlFileName: string;
-    xmlFile: string;
-    ecccZipFileByZipFileId: {
-      zipFileName: string;
-    };
-  };
-}
+// NOTE: This file should not be edited
+// see https://nextjs.org/docs/basic-features/typescript for more information.
