+++ ggircs-app/app/package.json
@@ -45,13 +45,13 @@
     "lightship": "^6.7.2",
     "lodash.debounce": "^4.0.8",
     "morgan": "^1.10.0",
-    "next": "10.2.3",
+    "next": "11.1.3",
     "pg": "^8.5.1",
     "postgraphile": "^4.12.3",
     "postgraphile-log-consola": "^1.0.1",
     "postgraphile-plugin-connection-filter": "^2.2.2",
     "react": "^17.0.2",
-    "react-bootstrap": "^1.5.2",
+    "react-bootstrap": "^2.1.1",
     "react-diff-view": "^2.4.8",
     "react-dom": "^17.0.2",
     "react-relay": "^10.1.2",
