+++ ggircs-app/app/pages/_app.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { QueryRenderer } from "react-relay";
 import NextApp from "next/app";
 import { NextRouter } from "next/router";
-import { PageComponent } from "next-env";
+import { PageComponent } from "types";
 import { getRequest } from "relay-runtime";
 import { createEnvironment } from "lib/relay-environment";
 import ErrorBoundary from "lib/error-boundary";
