+++ ggircs-app/app/pages/login-redirect.tsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Row, Col } from "react-bootstrap";
 import { graphql } from "react-relay";
 import { loginRedirectQueryResponse } from "loginRedirectQuery.graphql";
-import { PageComponentProps } from "next-env";
+import { PageComponentProps } from "types";
 import DefaultLayout from "components/Layout/DefaultLayout";
 import { NextRouter } from "next/router";
