+++ ggircs-app/app/pages/request-access-notice.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { graphql } from "react-relay";
 import { requestAccessNoticeQueryResponse } from "requestAccessNoticeQuery.graphql";
-import { PageComponentProps } from "next-env";
+import { PageComponentProps } from "types";
 import DefaultLayout from "components/Layout/DefaultLayout";
 import { PENDING_GGIRCS_USER } from "data/group-constants";
