+++ ggircs-app/app/server/schema.graphql
@@ -3179,7 +3179,6 @@ type Query implements Node {
   session: JwtToken
 }
 
-"""The table housing all report data, derived from eccc_xml_file table"""
 type Report implements Node {
   co2Bioc: BigFloat
 
@@ -3203,8 +3202,6 @@ type Report implements Node {
 
   """The timestamp recorded in SWRS when the report was last modified"""
   lastModifiedDate: Datetime
-
-  """Computed column returns the latest report ordered by zip_file_name"""
   latestSwrsReport: Report
 
   """The prepop report id"""
@@ -3281,7 +3278,6 @@ type Report implements Node {
   version: String
 }
 
-"""The table housing all attachments for reports"""
 type ReportAttachment implements Node {
   comment: String
